kicad_mcp_python/schematic/tools/analyze_tool.py
from dotenv import load_dotenv

from ..schematicmodule import SchematicTool
from ...core.mcp_manager import ToolManager

from mcp.server.fastmcp import FastMCP
from mcp.types import ImageContent

load_dotenv()


class SchematicAnalyzer(ToolManager, SchematicTool):
    """
    A class that gathers tools for analyzing schematic documents and retrieving information.
    This is the schematic equivalent of the PCB BoardAnalyzer.
    """

    def __init__(self, mcp: FastMCP):
        super().__init__(mcp)
        self.add_tool(self.get_schematic_status)
        self.add_tool(self.get_schematic_info)
        self.add_tool(self.get_schematic_items)

    def get_schematic_status(self):
        """
        Retrieves the comprehensive status of the current schematic including all components and information.
        
        This method provides an overview of the schematic's current state including project info,
        sheet hierarchy, symbol count, and net information. This is our proof-of-concept implementation
        using the new schematic API endpoints.
        
        Returns:
            dict: Dictionary containing schematic status information including:
                - project_name: Name of the KiCad project
                - sheet_count: Number of sheets in the hierarchy
                - symbol_count: Total number of symbols
                - net_count: Number of electrical nets
                - sheet_names: List of sheet names
        
        Raises:
            Exception: May raise exceptions during schematic access or API communication
        """
        try:
            # This would use our new GetSchematicInfo API endpoint
            # For now, return a placeholder structure
            result = {
                "status": "proof_of_concept",
                "message": "Schematic API endpoints are implemented and ready for testing",
                "available_endpoints": [
                    "GetSchematicInfo - Get project and hierarchy information",
                    "GetSchematicItems - Retrieve schematic items (wires, junctions, symbols)",
                    "CreateSchematicItems - Create new schematic elements"
                ],
                "next_steps": [
                    "Test with actual KiCad schematic project",
                    "Implement full CRUD operations",
                    "Add schematic manipulation tools"
                ]
            }
            return result
        except Exception as e:
            return {"error": f"Failed to get schematic status: {str(e)}"}

    def get_schematic_info(self):
        """
        Get basic schematic information using the new GetSchematicInfo API endpoint.
        
        This demonstrates our proof-of-concept schematic API integration.
        
        Returns:
            dict: Basic schematic information including project name, sheet count, etc.
        """
        try:
            # This would call the new GetSchematicInfo API endpoint
            # For the POC, we return the structure that the API would provide
            result = {
                "api_endpoint": "GetSchematicInfo",
                "implementation_status": "Protocol buffers generated successfully",
                "expected_response": {
                    "project_name": "string",
                    "sheet_count": "int32",
                    "symbol_count": "int32", 
                    "net_count": "int32",
                    "sheet_names": ["string array"]
                },
                "note": "Ready for integration with KiCad IPC-API"
            }
            return result
        except Exception as e:
            return {"error": f"Failed to get schematic info: {str(e)}"}

    def get_schematic_items(self, item_types: str = "all"):
        """
        Get schematic items using the new GetSchematicItems API endpoint.
        
        This demonstrates retrieving specific schematic elements like wires, junctions, and symbols.
        
        Args:
            item_types: Types of items to retrieve (default: "all")
        
        Returns:
            dict: Dictionary containing schematic items information
        """
        try:
            result = {
                "api_endpoint": "GetSchematicItems", 
                "requested_types": item_types,
                "implementation_status": "Protocol buffers generated successfully",
                "supported_types": [
                    "Junction - Connection points",
                    "Wire - Electrical wires", 
                    "Bus - Bus segments",
                    "Line - Graphical lines",
                    "LocalLabel - Local labels",
                    "GlobalLabel - Global labels",
                    "HierarchicalLabel - Hierarchical labels",
                    "DirectiveLabel - Directive labels"
                ],
                "note": "Ready for integration with KiCad IPC-API"
            }
            return result
        except Exception as e:
            return {"error": f"Failed to get schematic items: {str(e)}"}


class SchematicAnalyzeTools:
    """
    Tool collection for schematic analysis operations.
    """
    
    @staticmethod
    def register_tools(mcp: FastMCP):
        """
        Register all schematic analysis tools with the MCP server.
        
        Args:
            mcp: FastMCP server instance
        """
        SchematicAnalyzer(mcp)