pyproject.toml
[tool.poetry]
name = "kicad_mcp_python"
version = "0.1.0"
description = "MCP server utilizing KiCad's official IPC-API"
authors = ["anthonypark <anthonypark6904@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
kicad-python = {path = "kicad-python", develop = true}
python-dotenv = "^1.0.0"
mcp = "^1.9.3"
cairosvg = "^2.7"
pytest = "^8.3.4"
watchdog = "^6.0.0"
setuptools = "^80.9.0"
protobuf = "^5.29"
protoletariat = "^3.3.10"
mypy-protobuf = "^3.6.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"